+++ ui/components/TracePage/InputOutput.stories.tsx
@@ -1,0 +1,26 @@
+/*import type { Meta, StoryObj } from "@storybook/react";
+
+import { InputOutputContent, InputOutputContentProps } from "./TracePage";
+
+const meta: Meta<typeof InputOutputContent> = {
+    title: "TracePage/FString",
+    component: InputOutputContent,
+};
+
+export default meta;
+type Story = StoryObj<typeof InputOutputContent>;
+
+const testArgs: InputOutputContentProps = {
+    args: [0, 0],
+    // taken from a demo workflow- build this
+    // TODO get this with json parse and json stringify from my data :)
+    result: [0, 0]
+    records: myRecords,
+}
+
+// TODO consider exporting inputoutputcontent into its own file
+export const HelloWorld: Story = {
+    args: testArgs
+};
+
+*/
